cualuong

+++ BookStoreClone/ViewModel/BaoCaoViewModel.cs
@@ -120,7 +120,7 @@ namespace BookStoreClone.ViewModel
                 {
                     foreach (var b in _listCTHD)
                     {
-                        if (a.MaHD == b.MaHD && b.PhuongThuc == "Mua") { s += b.SoLuong; }
+                        if (a.MaHD == b.MaHD && b.PhuongThuc == "Mua") { s += (int)b.SoLuong; }
                     }
                 }
                 SeriesCollection[1].Values.Add(s);
@@ -145,18 +145,7 @@ namespace BookStoreClone.ViewModel
 
 
 
-            //var result = from e in DataProvider.Ins.DB.HoaDons
-            //                      join d in DataProvider.Ins.DB.Saches
 
-            //			 where (e.NgayBan.Value.Month == Thang && e.NgayBan.Value.Year == Nam)
-            //			 orderby e.TongTien descending
-            //			 select new
-            //			 {
-            //				 e.BaoCaoCongNo.
-            //				 imageName = (from soh in db.tblProductImages
-            //							  where soh.product_id == e.ProductId
-            //							  select new { soh.image_name }).Take(1)
-            //			 };
             Nam = 2020;
             ListCTBaoCaoCongNo = new ObservableCollection<CTBaoCaoCongNo>(DataProvider.Ins.DB.CTBaoCaoCongNoes);
             ListCTBaoCaoTon = new ObservableCollection<CTBaoCaoTon>(DataProvider.Ins.DB.CTBaoCaoTons);
